Metodo nuevo en modelo de precio. Todo en controlador de precio. Controlador de precio actualizado. Vistas de precio terminadas

+++ producto 3/web/database/migrations/2024_11_24_053525_create_precio_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('id_precio');
             $table->unsignedBigInteger('id_vehiculo');
             $table->unsignedBigInteger('id_hotel');
-            $table->bigInteger('precio');
+            $table->decimal('precio');
 
             $table->foreign('id_vehiculo')->references('id_vehiculo')->on('p3_transfer_vehiculo');
             $table->foreign('id_hotel')->references( 'id_hotel')->on('p3_transfer_hotel');
